requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
altair==5.5.0
anyio==4.8.0
attrs==25.1.0
baml-py==0.89.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.112.2
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.16
langchain-community==0.3.16
langchain-core==0.3.32
langchain-openai==0.3.2
langchain-text-splitters==0.3.5
langgraph==0.2.68
langgraph-checkpoint==2.0.10
langgraph-checkpoint-sqlite==2.0.3
langgraph-sdk==0.1.51
langsmith==0.3.2
markdown-it-py==3.0.0
marker-pdf==0.2.6
MarkupSafe==3.0.2
marshmallow==3.26.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
narwhals==1.24.0
networkx==3.4.2
nest-asyncio==1.6.0
numpy==1.26.4
onnxruntime==1.20.1
openai==1.60.2
opencv-python==4.11.0.86
orjson==3.10.15
pdftext==0.3.7
pillow==10.4.0
propcache==0.2.1
pyarrow==19.0.0
pydantic-settings==2.7.1
pydeck==0.9.1
PyMySQL==1.1.1
pypdfium2==4.30.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.12.1
ray==2.42.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==72.1.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.37
streamlit==1.41.1
surya-ocr==0.4.5
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
texify==0.1.10
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
tqdm==4.67.1
transformers==4.36.2
typing-inspect==0.9.0
urllib3==2.3.0
uvicorn==0.32.1
watchdog==6.0.0
wheel==0.44.0
yarl==1.18.3
zstandard==0.23.0
